Weights v2 but need to rewrite eggs

+++ Java/src/Hex.java
@@ -17,10 +17,10 @@ public class Hex {
     int myAnts; // the amount of your ants on this cell
     int oppAnts; // the amount of opponent ants on this cell
 
-    double value;
+    int value;
     int distanceToMyBase;
 
-    public Hex(int index, int type, int initialResources, int neigh0, int neigh1, int neigh2, int neigh3, int neigh4, int neigh5, int resources, int myAnts, int oppAnts, double value) {
+    public Hex(int index, int type, int initialResources, int neigh0, int neigh1, int neigh2, int neigh3, int neigh4, int neigh5, int resources, int myAnts, int oppAnts, int value) {
         this.index = index;
         this.type = type;
         this.initialResources = initialResources;
@@ -136,11 +136,11 @@ public class Hex {
         this.oppAnts = oppAnts;
     }
 
-    public double getValue() {
+    public int getValue() {
         return value;
     }
 
-    public void setValue(double value) {
+    public void setValue(int value) {
         this.value = value;
     }
